Fix energy values overflowing on servers (fixes #2)
Minecraft sends container properties as 16-bit shorts instead of 32-bit ints. I changed the IIntArrays to break the energy value into two 16-bit chunks, which are recombined on the client to produce the correct value. Also changed the machine IIntArrays to use 0 and 1 for energy, so it is consistent with the battery box.