New Textures
I noticed that the textures included in this commit were previously temporary. This mod is excellent so I wanted to bring forth some new textures for the temp blocks which are both animated when powered on (to clearly show to the player that the machine is running) and within the theme of the other assets of the mod.

The model files have also been modified to support these new textures.